# 셀프진단 프롬프트 개선 및 서비스 설정 업데이트
## 프롬프트 개선
- selfcheck_individual.md: 개별 항목 분석 프롬프트 강화
- selfcheck_similarity.md: 유사과제 검토 프롬프트 개선
- kca-certification_prompt.md: KCA 인증 프롬프트 업데이트
- questions_prompt.md: 질문 생성 프롬프트 확장

## 프론트엔드 개선
- analysis-progress.tsx: 분석 진행 UI 개선
- selfcheck/page.tsx: 셀프진단 페이지 수정

## 백엔드 개선
- selfcheck.py: 셀프진단 API 라우트 확장
- document_selector_service.py: 문서 선택 서비스 수정
- prompt_generator_service.py: 프롬프트 생성 서비스 개선

## 서비스 설정
- docs/service/*: LLM 서비스 설정 파일 업데이트
- exaone-deep 서비스 파일 제거 (통합)

+++ backend/services/document_selector_service.py
@@ -125,7 +125,7 @@ class DocumentSelectorService:
         if not content:
             return ""
 
-        max_chars = max_tokens * self.CHARS_PER_TOKEN
+        max_chars = int(max_tokens * self.CHARS_PER_TOKEN)
 
         if len(content) <= max_chars:
             # 전체 내용이 제한보다 작으면 그대로 반환
@@ -288,7 +288,7 @@ class DocumentSelectorService:
 
             all_chunks = []
             seen_chunk_ids = set()
-            max_chars = max_tokens_total * self.CHARS_PER_TOKEN
+            max_chars = int(max_tokens_total * self.CHARS_PER_TOKEN)
 
             # 각 쿼리로 검색하여 다양한 관점의 청크 수집
             chunks_per_query = max(5, top_k // len(search_queries))
